update fp types

+++ sahi/models/yolov8trt.py
@@ -132,7 +132,7 @@ class Yolov8TrtDetectionModel(DetectionModel):
 
         input_image = input_image / 255.0
         input_image = input_image.transpose(2, 0, 1)
-        image_tensor = input_image[np.newaxis, :, :, :].astype(np.float32)
+        image_tensor = input_image[np.newaxis, :, :, :].astype(np.float16)
 
         return image_tensor
